Refactor Library Components to Use Modular Structure
