+++ autogpt_platform/frontend/src/app/(platform)/library/components/LibraryNotificationCard/library-notification-card.tsx
