+++ autogpt_platform/frontend/src/app/(platform)/library/components/state-provider.tsx
